Adding parent, admin and educator controllers

+++ backend/src/main/kotlin/com/ablx/daycare/backend/enum/Role.kt
@@ -1,6 +1,6 @@
 package com.ablx.daycare.backend.enum
 
-enum internal class Role(val value: Int) {
+enum internal class Role(val value: Long) {
     EDUCATOR(1),
     ADMIN(2),
     PARENT(3)
